requirements.txt
streamlit
pandas
datetime
numpy
gdelt
scikit-learn
wordcloud
matplotlib.pyplot